civicflow/backend/pyproject.toml
[project]
name = "civicflow-backend"
version = "0.1.0"
description = "CivicFlow Flask API backend"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-cors>=4.0",
    "pymongo>=4.6",
    "pyjwt>=2.8",
    "fpdf2>=2.7",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "bcrypt>=5.0.0",
    "sarvamai>=0.1.26",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.uv]
package = false
